激活界面提交9

+++ C/aesCode.h
@@ -7,8 +7,8 @@
 #include <stdint.h>
 #include <string.h>
 
-bool EncryptDataToCipherTxt(uint8_t *orign, uint8_t *result, uint16_t length);
+bool EncryptDataToCipherTxt(uint8_t *orign, uint8_t *result, int length);
 
-bool DecryptCipherTxtToData(uint8_t *orign, uint8_t *result, uint16_t length);
+bool DecryptCipherTxtToData(uint8_t *orign, uint8_t *result, int length);
 
 #endif
